rhtap: update Dockerfile to obtain a clean git hash
The cachi2 pipelinerun task in RHTAP modifies the Dockerfile to inject
the cached dependency information into the build. This causes the git
version - which is injected into the CLI version output - to include a
`-dirty` suffix.

The modification in this commit stashes those changes, and then gets the
git hash, storing it in the `GIT_VERSION` environment variable, then
pops those changes back off of the stash and runs the build as before.

The commit also contains some optimizations so that most of the `RUN`
commands are consolidated into one.

Signed-off-by: Lance Ball <lball@redhat.com>